refactor: 노트 관련 모달 리팩토링
- useModalPortal 추가
